refactor(user-management): Fix error handling

+++ user-management/src/controllers/UserController.js
@@ -1,4 +1,5 @@
 import User from "../repositories/UserRepository.js";
+import utils from "../utils.js";
 
 export const getUsers = async (req, res) => {
   try {
@@ -7,7 +8,7 @@ export const getUsers = async (req, res) => {
     const users = await User.getUsers({ filter, select });
     res.status(200).json(users);
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    return utils.handleError(res, error);
   }
 };
 
@@ -15,11 +16,11 @@ export const getUserById = async (req, res) => {
   try {
     const user = await User.getUserById(req.params.id);
     if (!user) {
-      return res.status(404).json({ message: "User not found" });
+      throw { code: 404, message: "User not found" };
     }
     res.status(200).json(user);
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    return utils.handleError(res, error);
   }
 };
 
@@ -27,25 +28,29 @@ export const updateUserById = async (req, res) => {
   try {
     const currentUser = await User.getUserById(req.params.id);
     if (!currentUser) {
-      return res.status(404).json({ message: "User not found" });
+      throw { code: 404, message: "User not found" };
     }
     const userData = { ...currentUser, ...req.body };
     if (!await User.updateUserById(userData)) {
-      return res.status(400).json({ message: "Update failed" });
+      throw { code: 400, message: "Update failed" }
     }
     res.status(200).json(await User.getUserById(userData.id));
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    return utils.handleError(res, error);
   }
 };
 
 export const deleteUserById = async (req, res) => {
   try {
+    const currentUser = await User.getUserById(req.params.id);
+    if (!currentUser) {
+      throw { code: 404, message: "User not found" };
+    }
     if (!await User.deleteUserById(req.params.id)) {
-      return res.status(404).json({ message: "User not found" });
+      throw { code: 400, message: "Delete failed" };
     }
     res.status(204).json({});
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    return utils.handleError(res, error);
   }
 };
